requirements.txt
# QDT Dashboard - Python Dependencies
# Install with: pip install -r requirements.txt

pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
sendgrid>=6.9.0
python-dotenv>=1.0.0
joblib>=1.3.0
tqdm>=4.65.0
plotly>=5.15.0
pyarrow>=14.0.0
flask>=2.3.0
flask-cors>=4.0.0

